match Stockfighter's language on incoming orders
The real Stockfighter API recognizes both `stock` and `symbol` as keys
on incoming orders (and possibly other, undiscovered synonyms). However,
the disorder book would recognize only the `symbol` form and emit a
cryptic error (the result of an assertion failure) if the client made
use of the other form. Sadly, `stock` is the form given in the
documentation.

With this change, the disorder book will recognize either form. If both
are present, `stock` silently takes priority.

Fixes #1